% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{add_perc}
\alias{add_perc}
\title{add_perc}
\usage{
add_perc(data)
}
\arguments{
\item{data}{A data frame that has a date or similar variable as its first
column, followed by the levels of some categorical variable that makes up a
total, which is the last column. Typically a data frame created by the
\code{\link{vardistr_amnt}} function.}
}
\value{
the same data frame plus a column starting with "perc" for each of
  the existing variables. These added variables will add up to 100% if they
  are summed up row wise.
}
\description{
adds percentage distribution columns to a summarised data frame
}
\examples{
x <- add_perc(df_vda)

}
