% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_perc_cpp}
\alias{add_perc_cpp}
\title{add_perc_cpp}
\usage{
add_perc_cpp(df, start_col)
}
\arguments{
\item{data}{A data frame that has a date or similar variable as its first
column, followed by the levels of some categorical variable that makes up a
total, which is the last column. Typically a data frame created by the
\code{\link{vardistr_amnt}} function.
@param start_col start start summing across from this column number}
}
\value{
the same data frame plus a column starting with "perc" for each of
  the existing variables. These added variables will add up to 100% if they
  are summed up row wise.
}
\description{
The c++ version of \code{\link{add_perc}} function.
  Adds percentage distribution columns to a summarised data frame
}
\examples{
x <- add_perc_cpp(df, 2)

}
