% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is_null_string}
\alias{is_null_string}
\title{is_null_string}
\usage{
is_null_string(w)
}
\arguments{
\item{w}{a string. Either "NULL" or any other "string".}
}
\value{
TRUE or FALSE. Returns TRUE if the string is "NULL"
}
\note{
for use when writing functions with dplyr where function needs string as input.
}
\examples{
is_null_string("NULL")
is_null_string("Something_else")
}
