% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{last_day}
\alias{last_day}
\title{last_day}
\usage{
last_day(date)
}
\arguments{
\item{date}{}
}
\value{
last day of the month
}
\description{
calculate last day of the month given a date.
}
\examples{
last_day(date = as_date("2017-01-05"))

}
