% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_functions.R
\name{lead_flag}
\alias{lead_flag}
\title{lead_flag}
\usage{
lead_flag(data, outcome_period, flag)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary} format}

\item{outcome_period}{period to look ahead in terms of event (flag). E.g. for
12 month PD this will be a loss identification period (LIP) of 12.}

\item{flag}{the flag or event (could be default, settlement, write-off or any other permanent status)}
}
\value{
the input data frame with a parameter called lead_flag
}
\description{
creates a lead variable for a flag based on the outcome period
  (or loss_id_period in the case of PD calculations)
}
\examples{
# creates lead flag for calculating PD:
dft <- lead_flag(df, 12, "default_flag")

}
