% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{make_orig}
\alias{make_orig}
\title{make_orig}
\usage{
make_orig(data, use_period = FALSE, period = loan_period,
  period_number = 1)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard data
dictionary format}

\item{use_period}{either TRUE or FALSE. Default is FALSE, uses
pointintime_month == orig_month to filter. If TRUE then uses period and
period_number specified to filter.}

\item{period}{only specified if use_period = TRUE. Either fpd_period or
loan_period. used in conjunction with period_number to filter to unique
contract level.}

\item{period_number}{the loan_period or fpd_period to use when filtering.
usually 1.}
}
\value{
a loan level data set (should be unique on contract_key). for use in
  \code{\link{vardistr_amnt}} and \code{\link{vardistr_amnt}}
  vardistr_amnt functions.
}
\description{
condences monthly performance data set to origination level
}
\examples{

df_orig_test <- make_orig(df)
df_orig_test <- make_orig(df, use_period = TRUE, period = loan_period, period_number = 1)
}
