% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{max_var}
\alias{max_var}
\title{maximum of a variable in a data frame}
\usage{
max_var(data, var)
}
\arguments{
\item{data}{any data frame}

\item{var}{the name of a variable in the data frame upon which a maximum can be computed.}
}
\value{
the maximum of the variable
}
\description{
max_var
}
\examples{
max_month(df, "fpd_month")

}
