% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_functions.R
\name{over_booksize}
\alias{over_booksize}
\title{Overview of this month's book size}
\usage{
over_booksize(data, date)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary} format}

\item{date}{the latest point in time month of loan data available}
}
\value{
a list containing the booksize number for this month, last month
  and the % change in value from last month.
}
\description{
over_booksize
}
\examples{
#' dfbk <- over_booksize(df, "2017-06-30")

}
