% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_functions.R
\name{over_vinta_fpd}
\alias{over_vinta_fpd}
\title{Overview calculation for vintage analysis based on first pay date}
\usage{
over_vinta_fpd(data, date, default_definition)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary} format}

\item{date}{the latest performance month available}

\item{default_definition}{the default definition applicable to the loan
portfolio}
}
\value{
a list containing default rate for this month and the % change in
  value from last month for period 4 and 12.
}
\description{
over_vinta_fpd
}
\examples{
dft <- over_vinta_fpd(df, "2017-08-31", 4)

}
