% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_functions.R
\name{select_variables_pd}
\alias{select_variables_pd}
\title{select_variables_pd}
\usage{
select_variables_pd(data, segmenter_level = 1, segmenter_vector)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary} format}

\item{segmenter_level}{1, 2 or 3. Default = 1. When 1 then no segmentation
will be done when doing vintage analysis. If 2, then segmentation will be
done by to the variable entered as var1. If 3, then segmentation will be
done by var1 and var2. It is very important to enter the correct value
here.}

\item{segmenter_vector}{segmenter variables used for PD calculation. A vector
of variable names as strings.}
}
\value{
the input data frame with only the necessary variables.
}
\description{
selects the necessary variables from the standard data set for
  pd calculations
}
\examples{
df_select <- select_variables_pd(df, c("fico_bin"), 2)
df_select <- select_variables_pd(df,  1)
}
