% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_functions.R
\name{select_variables_pd_tr}
\alias{select_variables_pd_tr}
\title{select_variables_pd_tr}
\usage{
select_variables_pd_tr(data, ...)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary} format}

\item{...}{segmenter variables used for PD calculation}
}
\value{
the input data frame with only the necessary variables.
}
\description{
selects the necessary variables from the standard data set for pd calculations
}
\examples{
df_select <- select_variables_pd_tr(df, "fico_bin")
df_select <- select_variables_pd_tr(df, fico_bin)
}
