% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{string_to_quote}
\alias{string_to_quote}
\title{string_to_quote}
\usage{
string_to_quote(string)
}
\arguments{
\item{string}{a character string that represents a variable or expression.
The input parameter to a function.}
}
\value{
returns the quoted version of the string.
}
\description{
convert a string to a quoted expression within a function
}
\note{
Only use inside a function. Variable should be unquoted via UQ() or !!
  for use.
}
\examples{
input <- string_to_quote(input) #where input is a parameter entered as string

}
