% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_rates.R
\name{vintalyse}
\alias{vintalyse}
\title{vintalyse}
\usage{
vintalyse(data, period_dim = "loan_period", month_dim = "orig_month")
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
dictionary} format}

\item{period_dim}{the name of the period dimension used to do the vintage analysis. Usually either
"loan_period" or "fpd_period". Note, must be input as a "string".}

\item{month_dim}{the name of the month dimension used to do the vintage analysis. Usually either
"orig_month" or "fpd_month". Note, must be input as a "string".}
}
\value{
A data set used for plotting default vintage analysis.
}
\description{
produces
}
\note{
to be generalised to work for any flag. not just default.
}
\examples{
default_summary <- vintalyse(df, "loan_period", "orig_month")

}
