% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_functions.R
\name{get_cohort_base}
\alias{get_cohort_base}
\alias{get_cohort_default}
\title{get_cohort_base}
\usage{
get_cohort_base(data, default_definition, group_by_vector)

get_cohort_default(data, default_definition, group_by_vector)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary} format}

\item{default_definition}{the default definition applied to the loan
portfolio}

\item{group_by_vector}{Vector with string names of variables to be grouped
by. Created in pd function.}
}
\value{
A data frame containing the sum of closing balance by
  pointintime_month, months_arrears smaller than the default definition, and
  the segmenter variables used.
}
\description{
Calculates the total balance by arrears and date for each
  segmenter variables.
}
\section{Functions}{
\itemize{
\item \code{get_cohort_default}: Calculates the total balance by arrears and date for each
segmenter variables where the event (e.g. default) has occured.
}}

\note{
The cohort_base values essentially form the 'base' when doing PD
  calculations. cohort_base and cohort_default form the
}
\examples{
get_cohort_base(df, 3, c("months_arrears", "pointintime_month"))

}
