% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{vardistr_amnt}
\alias{vardistr_amnt}
\alias{vardistr_perc}
\title{vardistr_perc}
\usage{
vardistr_amnt(data, var, weight, datedim, ...)

vardistr_perc(data, var, weight, datedim, ...)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary}  format}

\item{var}{the variable by which you wish to show the monthly distribution.
Must be a "string".}

\item{weight}{either "NULL" or a weight variable such as "loan_amount" or
"closingbalance". When "NULL" the weight is count of contracts. Note, must
be entered as a "string".}

\item{datedim}{the date field by which you wish to show the distribution.
Usually "orig_month" or "fpd_month". "pointintime_month" is used when point
in time distributions want to be done over time (book analysis). Must be
entered as a "string".}

\item{...}{parameters used in \code{\link{make_orig}} function.}
}
\value{
a data frame with unique datedim in first column. other columns are
  the various levels of the var parameter. There is also a Total variable
  that contains the total of the weight parameter.
}
\description{
turns loan level data to monthly distributions
}
\section{Functions}{
\itemize{
\item \code{vardistr_amnt}: This only returns the amounts in each var level and the Total
}}

\examples{
df_vdp_test <- vardistr_perc(df, "fico_bin", "loan_amount", "orig_month", use_period = TRUE, period = fpd_period)
df_vdp_test <- vardistr_perc(df, "fico_bin", "NULL", "fpd_month")

df_vda <- vardistr_amnt(df, "fico_bin", "loan_amount", "orig_month", use_period = TRUE, period=fpd_period)
df_vda <- vardistr_amnt(df, "fico_bin", "loan_amount", "orig_month")

x <- vardistr_perc(df, "disclosure", "net_advance", "pointintime_month")
x <- vardistr_perc(df, "fico_bin", "closing_balance", "pointintime_month")

}
