% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_rates.R
\name{vintalyse}
\alias{vintalyse}
\title{vintalyse}
\usage{
vintalyse(data, ...)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary}  format}

\item{...}{date parameters to group by. Usually either loan_period,
orig_month or fpd_period, fpd_month.}
}
\value{
A data set used for plotting default vintage analysis.
}
\description{
produces
}
\note{
to be generalised to work for any flag. not just default.
}
\examples{
default_summary <- vintalyse(df, loan_period, orig_month)

}
