% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_rates.R
\name{vintalyse}
\alias{vintalyse}
\title{vintalyse}
\usage{
vintalyse(data, period_dim = "loan_period", month_dim = "orig_month",
  segmenter_level = 1, var1, var2)
}
\arguments{
\item{data}{a monthly loan performance level data frame in standard
\href{https://github.com/TheProfitTable/masterlibrary/blob/master/tpt_credit_datadictionary.Rmd}{data
 dictionary} format}

\item{period_dim}{the name of the period dimension used to do the vintage
analysis. Usually either "loan_period" or "fpd_period". Note, must be input
as a "string".}

\item{month_dim}{the name of the month dimension used to do the vintage
analysis. Usually either "orig_month" or "fpd_month". Note, must be input
as a "string".}

\item{segmenter_level}{1, 2 or 3. Default = 1. When 1 then no
segmentation will be done when doing vintage analysis. If 2, then
segmentation will be done by to the variable entered as var1. If
3, then segmentation will be done by var1 and var2.}

\item{var1}{the main variable to segment the vintage analysis by. Must be provided
as "string" and must be a categorical variable. See Note below.}

\item{var2}{if a second level of segmentation is required, this is the second level.}
}
\value{
A data set used for plotting default vintage analysis.
}
\description{
produces
}
\note{
Each level of the var1 and var2 pararmeters need to have sufficient volume to
  produce meaningful results. When creating a plot for the vintage analysis,
  the var1 and var2 parameters need to be either filtered upon or used to create a
  grid.
}
\examples{
default_summary <- vintalyse(df, "loan_period", "orig_month")
default_summary_var1 <- vintalyse(df, "loan_period", "orig_month", segmenter_level = 3, var1 = "fico_bin")
default_summary_var2 <- vintalyse(df, "loan_period", "orig_month", segmenter_level = 3, var1 = "fico_bin", "occpy_sts")
}
